Basic sync features completed

+++ lib/features/notes/data/repositories/notes_repository.dart
@@ -7,6 +7,8 @@ import 'package:dairy_app/features/notes/core/failures/failure.dart';
 import 'package:dairy_app/features/notes/domain/entities/notes.dart';
 import 'package:dairy_app/features/notes/domain/repositories/notes_repository.dart';
 import 'package:dartz/dartz.dart';
+import 'package:path/path.dart' as p;
+import 'package:crypto/crypto.dart';
 
 final log = printer("NotesRepository");
 
@@ -39,22 +41,34 @@ class NotesRepository implements INotesRepository {
 
   @override
   Future<Either<NotesFailure, void>> saveNote(
-      Map<String, dynamic> noteMap) async {
+    Map<String, dynamic> noteMap, {
+    bool dontModifyAnyParameters = false,
+  }) async {
     try {
-      List<NoteAsset> allNoteAssets = noteMap["asset_dependencies"];
-      List<String> usedNoteAssets = _parseAssets(noteMap["body"]);
+      if (dontModifyAnyParameters == false) {
+        List<NoteAsset> allNoteAssets = noteMap["asset_dependencies"];
+        List<String> usedNoteAssets = _parseAssets(noteMap["body"]);
 
-      print("all assets = $allNoteAssets");
-      print("used assets = $usedNoteAssets");
-
-      for (var noteAsset in allNoteAssets) {
-        if (!usedNoteAssets.contains(noteAsset.assetPath)) {
-          notesLocalDataSource.deleteFile(noteAsset.assetPath);
+        for (var noteAsset in allNoteAssets) {
+          if (!usedNoteAssets.contains(noteAsset.assetPath)) {
+            notesLocalDataSource.deleteFile(noteAsset.assetPath);
+          }
         }
+
+        noteMap["asset_dependencies"].removeWhere(
+            (noteAsset) => !usedNoteAssets.contains(noteAsset.assetPath));
+
+        // calculate note hash
+        String noteBodyWithAssetPathsRemoved =
+            _replaceAssetPathsByAssetNames(noteMap["body"]);
+
+        var _hash = _generateHash(noteMap["title"] +
+            noteMap["created_at"].toString() +
+            noteBodyWithAssetPathsRemoved);
+
+        noteMap["hash"] = _hash;
       }
 
-      noteMap["asset_dependencies"].removeWhere(
-          (noteAsset) => !usedNoteAssets.contains(noteAsset.assetPath));
       await notesLocalDataSource.saveNote(noteMap);
       return const Right(null);
     } catch (e) {
@@ -70,10 +84,6 @@ class NotesRepository implements INotesRepository {
       List<NoteAsset> allNoteAssets = noteMap["asset_dependencies"];
       List<String> usedNoteAssets = _parseAssets(noteMap["body"]);
 
-      print("UPdating note");
-      print("all note assets = $allNoteAssets");
-      print("used assets = $usedNoteAssets");
-
       for (var noteAsset in allNoteAssets) {
         if (!usedNoteAssets.contains(noteAsset.assetPath)) {
           notesLocalDataSource.deleteFile(noteAsset.assetPath);
@@ -82,6 +92,16 @@ class NotesRepository implements INotesRepository {
 
       noteMap["asset_dependencies"].removeWhere(
           (noteAsset) => !usedNoteAssets.contains(noteAsset.assetPath));
+
+      // calculate note hash
+      String noteBodyWithAssetPathsRemoved =
+          _replaceAssetPathsByAssetNames(noteMap["body"]);
+
+      var _hash = _generateHash(noteMap["title"] +
+          noteMap["created_at"].toString() +
+          noteBodyWithAssetPathsRemoved);
+
+      noteMap["hash"] = _hash;
 
       await notesLocalDataSource.updateNote(noteMap);
       return const Right(null);
@@ -103,10 +123,12 @@ class NotesRepository implements INotesRepository {
   }
 
   @override
-  Future<Either<NotesFailure, void>> deleteNotes(List<String> noteList) async {
+  Future<Either<NotesFailure, void>> deleteNotes(List<String> noteList,
+      {bool hardDeletion = false}) async {
     try {
       for (var noteId in noteList) {
-        await notesLocalDataSource.deleteNote(noteId);
+        await notesLocalDataSource.deleteNote(noteId,
+            hardDeletion: hardDeletion);
       }
       return const Right(null);
     } catch (e) {
@@ -115,8 +137,54 @@ class NotesRepository implements INotesRepository {
     }
   }
 
-  // utility functions
+  @override
+  Future<Either<NotesFailure, List<String>>> getAllNoteIds() async {
+    try {
+      var noteIdList = await notesLocalDataSource.getAllNoteIds();
+      return Right(noteIdList);
+    } catch (e) {
+      log.e(e);
+      return Left(NotesFailure.unknownError());
+    }
+  }
 
+  @override
+  Future<Either<NotesFailure, List<Map<String, dynamic>>>>
+      generateNotesIndex() async {
+    try {
+      var notesIndex = await notesLocalDataSource.getNotesIndex();
+      return Right(notesIndex);
+    } catch (e) {
+      log.e(e);
+      return Left(NotesFailure.unknownError());
+    }
+  }
+
+  @override
+  String replaceOldAssetPathsWithNewAssetPaths(
+      String noteBody, Map<String, dynamic> assetPathMap) {
+    var noteBodyMap = jsonDecode(noteBody);
+
+    for (var noteElement in noteBodyMap) {
+      if (noteElement.containsKey("insert") &&
+          noteElement["insert"].runtimeType != String) {
+        // replace paths from assetPathMap for image and video
+        if (noteElement["insert"].containsKey("image")) {
+          noteElement["insert"]["image"] =
+              assetPathMap[p.basename(noteElement["insert"]["image"])];
+        } else if (noteElement["insert"].containsKey("video")) {
+          noteElement["insert"]["video"] =
+              assetPathMap[p.basename(noteElement["insert"]["video"])];
+        }
+      }
+    }
+
+    return jsonEncode(noteBodyMap);
+  }
+
+  //* utility functions
+
+  /// Used to extract assets from note body
   List<String> _parseAssets(String noteBody) {
     var noteBodyMap = jsonDecode(noteBody);
     List<String> noteAssets = [];
@@ -132,5 +200,33 @@ class NotesRepository implements INotesRepository {
     }
 
     return noteAssets;
+  }
+
+  /// used to replace absolute paths of assets by their asset names, so that hash values
+  /// are same in all devices
+  String _replaceAssetPathsByAssetNames(String noteBody) {
+    var noteBodyMap = jsonDecode(noteBody);
+
+    for (Map<String, dynamic> noteElement in noteBodyMap) {
+      if (noteElement.containsKey("insert") &&
+          noteElement["insert"].runtimeType != String) {
+        // check for image and video types and replace their absolute paths with file names
+        if (noteElement["insert"].containsKey("image")) {
+          noteElement["insert"]["image"] =
+              p.basename(noteElement["insert"]["image"]);
+        } else if (noteElement["insert"].containsKey("video")) {
+          noteElement["insert"]["video"] =
+              p.basename(noteElement["insert"]["video"]);
+        }
+      }
+    }
+
+    return jsonEncode(noteBodyMap);
+  }
+
+  String _generateHash(String text) {
+    var bytes = utf8.encode(text);
+    var digest = sha1.convert(bytes);
+    return digest.toString();
   }
 }
